embed_docker/pyproject.toml
[tool.poetry]
name = "embedder-service"
version = "0.1.0"
description = "Semantic search based on embeddings"
authors = ["thingumajig <artifex.i@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
openpyxl = "^3.0.10"
Pillow = "9.1.0"
tensorflow-text = "2.4.3"
tensorflow_hub = "0.12.0"
pandas = "1.2.5"
tqdm = "^4.64.0"
ruclip = "0.0.1"


fastapi = ">= 0.68.0,<0.69.0"
pydantic = ">=1.8.0,<2.0.0"
uvicorn = ">=0.15.0,<0.16.0"
python-multipart = "0.0.5"
protobuf = "3.20.*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
